Refactor reading of books and ledger-cli invocation
