+++ largo/balance_sheet.py
@@ -1,8 +1,9 @@
 import subprocess
 from largo.project import Project
+from largo.ledger_invoke import LedgerInvoke
 
 
-class BalanceSheet:
+class BalanceSheet(LedgerInvoke):
     def __init__(self, project: Project):
         self.project = project
 
@@ -18,15 +19,3 @@ class BalanceSheet:
             arguments.extend(settings.default_options)
 
         return arguments
-
-    def read_book(self, year: int) -> bytes:
-        book_path = self.project.book(year)
-
-        with open(book_path, mode='rb') as book_file:
-            book = book_file.read()
-
-        return book
-
-    def build(self, year: int) -> subprocess.CompletedProcess:
-        book = self.read_book(year)
-        return subprocess.run(self.command_arguments, input=book)
