+++ largo/profit_loss.py
@@ -1,8 +1,9 @@
 from largo.project import Project
+from largo.ledger_invoke import LedgerInvoke
 import subprocess
 
 
-class ProfitLoss:
+class ProfitLoss(LedgerInvoke):
     def __init__(self, project: Project):
         self.project = project
 
@@ -13,15 +14,3 @@ class ProfitLoss:
                      self.project.account.income]
 
         return arguments
-
-    def read_book(self, year: int) -> bytes:
-        book_path = self.project.book(year)
-
-        with open(book_path, mode='rb') as book_file:
-            book = book_file.read()
-
-        return book
-
-    def build(self, year: int) -> subprocess.CompletedProcess:
-        book = self.read_book(year)
-        return subprocess.run(self.command_arguments, input=book)
